requirements.txt
Flask>=2.0
firebase-admin==6.5.0
python-dotenv==1.0.1
Flask-Cors==4.0.0
openai>=1.50.0
marshmallow==3.21.1
flask-jwt-extended>=4.6.0
